+++ retinazer/src/test/java/com/github/antag99/retinazer/HandleTest.java
@@ -1,6 +1,9 @@
 package com.github.antag99.retinazer;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 
 import org.junit.Test;
 
@@ -56,5 +59,16 @@ public class HandleTest {
         entity.destroy();
         engine.update();
         assertFalse(engine.getEntities().contains(entity.getEntity()));
+    }
+
+    @Test
+    public void testComponentsDestroyed() {
+        Engine engine = new Engine(new EngineConfig());
+        Handle entity = engine.createEntity().duplicate();
+        entity.create(FlagComponentA.class);
+        entity.destroy();
+        engine.update();
+        engine.createEntity();
+        assertNull(engine.getMapper(FlagComponentA.class).get(entity.getEntity()));
     }
 }