+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Version 0.1.2-SNAPSHOT
 - Made `Mask#getIndices(IntArray)` append elements to the array.
+- Fixed issue where destroying an entity didn't remove it's components
 
 # Version 0.1.1 (released 2015-08-21)
 - Updated to libgdx 1.6.5